requirements.txt
flask==3.0.3
requests==2.13.0
psycopg2-binary=2.9.9
